requirements.txt
# ── MQTT ─────────────────────────────────────────────────────
paho-mqtt==2.1.0

# ── Settings & Validation ───────────────────────────────────
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ── Utilities ───────────────────────────────────────────────
python-dotenv>=1.0.0
